Rename root script
